### 2025-05-26 #8277 @editorialbot
[PRE REVIEW]: genomesizeR: An R package for genome size prediction
**Submitting author:** @celine-mercier (<a href="http://orcid.org/0000-0002-4782-1530">Céline Mercier</a>)
**Repository:** https://github.com/ScionResearch/genomesizeR
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0.0001
