### 2025-05-27 3eab326
Updated references (added DOIs and changed brms ref)
