+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={27},
   pages={30--37},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.copbio.2013.09.012},
 }
 
 @Article{moran2002microbial,
@@ -16,7 +17,8 @@
   number={5},
   pages={583--586},
   year={2002},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/S0092-8674(02)00665-7},
 }
 
 @Article{raes2007prediction,
@@ -26,7 +28,8 @@
   volume={8},
   pages={1--11},
   year={2007},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1186/gb-2007-8-1-r10},
 }
 
 @Article{tyson2004community,
@@ -37,7 +40,8 @@
   number={6978},
   pages={37--43},
   year={2004},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi = {10.1038/nature02340},
 }
 
 @Article{vandenkoornhuyse2007active,
@@ -48,7 +52,8 @@
   number={43},
   pages={16970--16975},
   year={2007},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi = {10.1073/pnas.0705902104},
 }
 
 @Article{liu2023warmer,
@@ -59,7 +64,8 @@
   number={1},
   pages={e70},
   year={2023},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/imt2.70},
 }
 
 @Article{bates2015lme4,
@@ -81,6 +87,7 @@
   year = {2024},
   note = {R package version 0.6.2},
   url = {https://github.com/jknowles/mertools},
+  doi = {10.32614/CRAN.package.merTools},
 }
 
 @Manual{R2024,
@@ -98,6 +105,7 @@
   year = {2022},
   note = {R package version 1.0.17},
   url = {https://CRAN.R-project.org/package=doParallel},
+  doi = {10.32614/cran.package.doparallel},
 }
 
 @Article{phyloseq2013,
@@ -109,6 +117,7 @@
   number = {4},
   year = {2013},
   url = {http://dx.plos.org/10.1371/journal.pone.0061217},
+  doi = {10.1371/journal.pone.0061217},
 }
 
 @Article{schloss_introducting_2009,
@@ -121,7 +130,8 @@
 	doi = {10.1128/AEM.01541-09},
 	issn = {0099-2240},
 	URL = {https://aem.asm.org/content/75/23/7537},
-	journal = {Applied and Environmental Microbiology}
+	journal = {Applied and Environmental Microbiology},
+	doi = {10.1128/AEM.01541-09},
 }
 
 @Book{ggplot22016,
@@ -164,7 +174,8 @@
   month     =  jun,
   year      =  2023,
   copyright = "https://creativecommons.org/licenses/by/4.0",
-  language  = "en"
+  language  = "en",
+  doi = {10.1038/s41467-023-37937-4},
 }
 
 @Article{OLeary2016-kw,
@@ -195,7 +206,8 @@
   pages     = "D733--45",
   month     =  jan,
   year      =  2016,
-  language  = "en"
+  language  = "en",
+  doi = {10.1093/nar/gkv1189},
 }
 
 @Article{Dick2019-nl,
@@ -207,7 +219,8 @@
   volume    =  7,
   month     =  jul,
   year      =  2019,
-  copyright = "https://creativecommons.org/licenses/by/4.0/"
+  copyright = "https://creativecommons.org/licenses/by/4.0/",
+  doi = {10.3389/feart.2019.00180},
 }
 
[…]
