### 2025-07-25 #8277 @abhishektiwari
_My Pre-Review Checklist_

- [ ]  License - [OSI approved](https://opensource.org/licenses/alphabetical) software license
- [ ]  Substantial scholarly effort (Age of software/No. of commits, LOC/Existing Citations/Sufficiently Useful)

- [ ] A statement of need
- [ ] Community guidelines

- [ ] Paper compiles to PDF
- [ ] Paper's length/word count
- [ ] Reference is correctly rendered
- [ ]  No unused citations in `.bib` file
- [ ] DOI Check Passed - Check SKIP/ MISSING/INVALID DOIs
- [ ] Contribution and authorship - author list appropriate and complete
