### 2025-07-25 #8277 @abhishektiwari
Hi @celine-mercier I will be helping to edit this paper. I see current paper is length ~2000 words. Can you please bring down the word count for `paper.md`  ~1000 words?

Let me know if any questions.
