+++ paper/paper.md
@@ -14,7 +14,7 @@ authors:
     affiliation: '1'
   - name: Sean Husheer
     orcid: "0000-0002-0422-2386"
-    affiliation: "1, 2"
+    affiliation: "1"
   - name: Loretta Garrett
     orcid: "0000-0001-7687-6795"
     affiliation: '1'
@@ -22,33 +22,26 @@ authors:
     orcid: "0000-0002-1167-8699"
     affiliation: '1'
 affiliations:
-  - name: Scion Research, PO Box 3020, Rotorua 3010, New Zealand
+  - name: Scion Group, Bioeconomy Science Institute, Titokorangi Drive, Private Bag 3020, Rotorua 3046, New Zealand
     index: 1
-  - name: Havelock North Research Centre, The New Zealand Institute for Plant and Food Research Limited, Havelock North, New Zealand
-    index: 2
-date: 22 May 2025
+date: 04 August 2025
 bibliography: paper.bib
 
 ---
 
 # Summary
 
-The genome size of organisms present in an environment can provide many insights into evolutionary and ecological processes at play in that environment. The genomic revolution has enabled a rapid expansion of our knowledge of genomes in many living organisms, and most of that knowledge is classified and readily available in the databases of the National Center for Biotechnology Information (NCBI). The `genomesizeR` tool leverages the wealth of taxonomic and genomic information present in NCBI databases to infer the genome size of Archeae, Bacteria, or Eukaryote organisms identified at any taxonomic level. 
-This R package uses statistical modelling on data from the most up-to-date NCBI databases and provides three statistical methods for genome size prediction of a given taxon, or group of taxa. A straightforward 'weighted mean' method identifies the closest taxa with available genome size information in the taxonomic tree, and averages their genome sizes using weights based on taxonomic distance. A frequentist random effect model uses nested genus and family information to output genome size estimates. Finally, a third option provides predictions from a distributional Bayesian multilevel model which uses taxonomic information from genus all the way to superkingdom, therefore providing estimates and uncertainty bounds even for under-represented taxa.
+The genome size of organisms present in an environment can provide many insights into evolutionary and ecological processes at play in that environment. The genomic revolution has enabled a rapid expansion of our knowledge of genomes in many living organisms, and most of that knowledge is classified and readily available in the databases of the National Center for Biotechnology Information (NCBI). The `genomesizeR` tool leverages the wealth of taxonomic and genomic information present in NCBI databases to infer the genome size of Archeae, Bacteria, or Eukaryote organisms identified at any taxonomic level.
 
-All three methods use:
-
-  - A list of queries; a query being a taxon or a list of several taxa. The package was designed to make it easy to use with data coming from environmental DNA experiments, but works with any table of taxa.
-  - A reference database containing all the known genome sizes, built from the NCBI databases, with associated taxa, provided in an archive to download.
-  - A taxonomic tree structure as built by the NCBI, provided in the same archive.
+This R package provides three statistical methods for genome size prediction of a given taxon, or group of taxa. A straightforward 'weighted mean' method identifies the closest taxa with available genome size information in the taxonomic tree, and averages their genome sizes using weights based on taxonomic distance. A frequentist random effect model uses nested genus and family information to output genome size estimates. Finally, a third option provides predictions from a distributional Bayesian multilevel model which uses taxonomic information from genus all the way to superkingdom, therefore providing estimates and uncertainty bounds even for under-represented taxa.
 
 `genomesizeR` retrieves the taxonomic classification of input queries, estimates the genome size of each query, and provides 95% confidence intervals for each estimate. Several plotting functions are also provided to visualise the results.
 
 # Statement of need
 
[…]
