### 2025-08-03 #8277 @celine-mercier
Thank you @abhishektiwari , we have reduced the word count to ~1,000 words.
