### 2025-08-04 #8277 @abhishektiwari
Thanks @celine-mercier, 2 more tasks

- There are a few unused citation keys in `paper.bib` which are not referenced in the `paper.md` anymore? I guess they were originally referenced in the longer version of `paper.md`. Can you remove them?

```
biomformat2024
burkner2021brms
Dick2019-nl
dplyr2023
ggplot22016
knowles2024mertools
Labouyrie2023-yc
liu2023warmer
ncbitax2020
pbapply2023
phyloseq2013
R2024
raes2007prediction
schloss_introducting_2009
tyson2004community
weston2022doparallel
Yu2020-pa
```

- Please add contribution guideline and to mention/link to them in your README. You could for instance create a CONTRIBUTING.md file (see here for some examples: https://contributing.md/example/).

Once you have completed these, can you suggest a few potential reviewers? You can also share GitHub handles of potential reviewers. Please be sure not to @-tag your suggestions, to avoid unwanted GitHub notifications to potential reviewers. Please also read the our [conflict of interest](https://github.com/openjournals/joss/blob/main/COI.md) (COI) when recommending reviewers.
