### 2025-08-04 31d4933
Updated .bib file to remove unused references
