+++ paper/paper.bib
@@ -21,29 +21,6 @@
   doi = {10.1016/S0092-8674(02)00665-7},
 }
 
-@Article{raes2007prediction,
-  title={Prediction of effective genome size in metagenomic samples},
-  author={Raes, Jeroen and Korbel, Jan O and Lercher, Martin J and Von Mering, Christian and Bork, Peer},
-  journal={Genome biology},
-  volume={8},
-  pages={1--11},
-  year={2007},
-  publisher={Springer},
-  doi = {10.1186/gb-2007-8-1-r10},
-}
-
-@Article{tyson2004community,
-  title={Community structure and metabolism through reconstruction of microbial genomes from the environment},
-  author={Tyson, Gene W and Chapman, Jarrod and Hugenholtz, Philip and Allen, Eric E and Ram, Rachna J and Richardson, Paul M and Solovyev, Victor V and Rubin, Edward M and Rokhsar, Daniel S and Banfield, Jillian F},
-  journal={Nature},
-  volume={428},
-  number={6978},
-  pages={37--43},
-  year={2004},
-  publisher={Nature Publishing Group UK London},
-  doi = {10.1038/nature02340},
-}
-
 @Article{vandenkoornhuyse2007active,
   title={Active root-inhabiting microbes identified by rapid incorporation of plant-derived carbon into RNA},
   author={Vandenkoornhuyse, Philippe and Mah{\'e}, St{\'e}phane and Ineson, Philip and Staddon, Phil and Ostle, Nick and Cliquet, Jean-Bernard and Francez, Andr{\'e}-Jean and Fitter, Alastair H and Young, J Peter W},
@@ -56,18 +33,6 @@
   doi = {10.1073/pnas.0705902104},
 }
 
-@Article{liu2023warmer,
-  title={Warmer and drier ecosystems select for smaller bacterial genomes in global soils},
-  author={Liu, Hongwei and Zhang, Haiyang and Powell, Jeff and Delgado-Baquerizo, Manuel and Wang, Juntao and Singh, Brajesh},
-  journal={iMeta},
-  volume={2},
-  number={1},
-  pages={e70},
-  year={2023},
-  publisher={Wiley Online Library},
-  doi={10.1002/imt2.70},
-}
-
 @Article{bates2015lme4,
   title = {Fitting Linear Mixed-Effects Models Using {lme4}},
   author = {Douglas Bates and Martin M{\"a}chler and Ben Bolker and
@@ -78,104 +43,6 @@
   number = {1},
   pages = {1--48},
   doi = {10.18637/jss.v067.i01},
-}
-
-@Manual{knowles2024mertools,
-  title = {merTools: Tools for Analyzing Mixed Effect Regression
-    Models},
-  author = {Jared E. Knowles and Carl Frederick},
-  year = {2024},
-  note = {R package version 0.6.2},
-  url = {https://github.com/jknowles/mertools},
-  doi = {10.32614/CRAN.package.merTools},
-}
-
-@Manual{R2024,
-  title = {R: A Language and Environment for Statistical Computing},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2024},
-  url = {https://www.R-project.org/},
-}
-
-@Manual{weston2022doparallel,
-  title = {doParallel: Foreach Parallel Adaptor for the 'parallel' Package},
-  author = {Microsoft Corporation and Steve Weston},
-  year = {2022},
-  note = {R package version 1.0.17},
-  url = {https://CRAN.R-project.org/package=doParallel},
-  doi = {10.32614/cran.package.doparallel},
-}
-
[…]
