### 2025-08-04 a465aff
Added contributing guidelines and linked in Readme
