### 2025-08-04 53baad3
Fixed links to issues in contributing guidelines
