### 2025-08-05 #8277 @celine-mercier
Hi @abhishektiwari

Here's a list of potential reviewers:

Georg Rüppel	@ g-rppl	(Julia, R, ecology, radio-tracking, movement analysis, Bayesian statistics, probabilistic programming, literate programming, time series)

Adrien Taudière	@ adrientaudiere	(R, bioinformatics, statistics in ecology, metabarcoding)

BrandonEdwards	@ BrandonEdwards	(C/C++, Julia, Python, R	Ecology, Bayesian, Statistics)

Mohamed S. Sarhan	@ msabrysarhan	(Bash/Shell, Python, R, Microbiology, Ancient DNA, Microbial Ecology, Plant science, and Genomics)

ghislainv	@ ghislainv	(Python, R, Ecology Community ecology Conservation biology Bayesian statistics Tropical forest ecology Landscape ecology)

Thanks!
