### 2025-08-05 #8277 @abhishektiwari
👋 @msabrysarhan @g-rppl @ghislainv, would any of you be willing to review this submission for JOSS?  The submission under consideration is "genomesizeR: An R package for genome size prediction".

Based on your experience, we think you might be able to provide a great review of this submission.
Thanks
Abhi
