### 2025-08-06 #8277 @BrandonEdwards
> 👋 [@BrandonEdwards](https://github.com/BrandonEdwards) [@adrientaudiere](https://github.com/adrientaudiere) - would any of you be willing to review this submission for JOSS? The submission under consideration is "genomesizeR: An R package for genome size prediction".
>
> Given your experience and priori review for JOSS, I think you might be able to provide a great review of this submission.
>
> Abhi

Do you still need me as a reviewer? Happy to do so if yes.
