### 2025-09-11 ScionResearch/genomesizeR#1 @BrandonEdwards
Trouble installing, "ls-remote https://git.bioconductor.org/packages/ggtree RELEASE_3_21' had status 128"
Hi there! I am one of the reviewers for the JOSS submission associated with this R package (https://github.com/openjournals/joss-reviews/issues/8759#issuecomment-3220017844). My instructions are to create an issue in the target repo (i.e., this repo) if I run into any issues with the review checklist. I seem to be having issues installing this package following the instructions on the README.

When I run the command:
```remotes::install_github("ScionResearch/genomesizeR")```

I get the following error message:
`Using GitHub PAT from the git credential store.
Downloading GitHub repo ScionResearch/genomesizeR@HEAD
Error: Failed to install 'genomesizeR' from GitHub:
  Command failed (128)
In addition: Warning message:
In system(full, intern = TRUE, ignore.stderr = quiet) :
  running command '"C:\Users\B_EDWA~1\AppData\Local\Programs\Git\cmd\git.exe" ls-remote https://git.bioconductor.org/packages/ggtree RELEASE_3_21' had status 128`

I wasn't sure if it was some issue with a need to install ggtree, and so I attempted to install ggtree separately which worked fine, and I still get the same error as above.

Here is my sessionInfo():

```> sessionInfo()
R version 4.4.2 (2024-10-31 ucrt)
Platform: x86_64-w64-mingw32/x64
Running under: Windows 10 x64 (build 19045)

Matrix products: default

locale:
[1] LC_COLLATE=English_United States.utf8  LC_CTYPE=English_United States.utf8
[3] LC_MONETARY=English_United States.utf8 LC_NUMERIC=C
[5] LC_TIME=English_United States.utf8

time zone: America/Chicago
tzcode source: internal

attached base packages:

[1] BiocManager_1.30.26 compiler_4.4.2      tools_4.4.2         curl_6.2.0          remotes_2.5.0      ```

Please let me know if you need any more information!
