### 2025-09-14 ScionResearch/genomesizeR#1 @celine-mercier
Hi! Thanks for reviewing the paper.

Could you try those 4 alternatives:

```
remotes::install_github("https://github.com/ScionResearch/genomesizeR")

remotes::install_git("https://github.com/ScionResearch/genomesizeR")

install.packages("pak")
pak::pkg_install("git::https://github.com/ScionResearch/genomesizeR")

install.packages("devtools")
devtools::install_github("ScionResearch/genomesizeR")
```

Does any of those 4 options work for you? Thank you
