### 2025-09-17 ScionResearch/genomesizeR#1 @BrandonEdwards
Thank you! The first two options produced the same error, but I did have success installing the package with the third option. I am very unclear if it is something with my machine, or corporate firewall related things (I have been having issues with other R packages in that way), but option #3 seemed to work no problem. I will continue with my review.
