### 2025-09-18 ScionResearch/genomesizeR#1 @celine-mercier
Each of those options uses a slightly different underlying system, and I guess it's a matter of finding which one is allowed by the corporate setup :)

Thank you for testing, I will update the installation instructions to include all options.
