### 2025-09-20 #8759 @msabrysarhan
Dear @abhishektiwari , dear @celine-mercier ,
# Installation
I tested the package in my computer with windows 11 operating sys. It worked smoothly without any issues.
# Documentation
- Downloading and testing the reference data and the example csv file was not straight forward to me. It was not clear to me how to format the input file. I'd have preferred to have the function `estimate_genome_size()` accept taxids or vector of multiple taxids. However, this is very optional.
- In the abstract, the authors mentioned "Several plotting functions ...". I only saw three plotting functions. I would recommend to remove "Several".
# Testing
I tried to run the function with other TaxIDs at species, genus and class level, but kept getting the same values.
Here is the taxa I tested:
TaxIDs: 162481, 1799530, 418497
Could you please double check for that? I expected to get still estimate at higher taxonomic levels.

Otherwise, I think it is a very interesting tool. I personally would need to with long-read genome assemblies for fungi that have no reference genomes available.
