### 2025-09-24 22b2013
Added more options for installation. Closes #1
