### 2025-09-24 3d73eb0
Paper: Bayesian method now described in more detailed and brms cited
