+++ paper/paper.bib
@@ -77,5 +77,16 @@
   doi = {10.1093/nar/gkv1189},
 }
 
-
+@article{burkner2021brms,
+  title = {Bayesian Item Response Modeling in {R} with {brms} and
+    {Stan}},
+  author = {Paul-Christian Bürkner},
+  journal = {Journal of Statistical Software},
+  year = {2021},
+  volume = {100},
+  number = {5},
+  pages = {1--54},
+  doi = {10.18637/jss.v100.i05},
+  encoding = {UTF-8},
+}
 
