### 2025-10-01 b7d9402
Move tutorial to Get Started page
