+++ vignettes/genomesizeR.Rmd
@@ -1,16 +1,41 @@
 ---
-title: "Example"
+title: "Get started"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Example}
+  %\VignetteIndexEntry{getstarted}
   %\VignetteEngine{knitr::rmarkdown}
   \usepackage[utf8]{inputenc}
 ---
 
-#### Example using default bayesian method on example file
+#### Simple tutorial using default bayesian method on example file
 
-This example data is a subset of the dataset from @Labouyrie2023-yc.
+If not already done, download the archive containing the reference databases and the bayesian models from `zenodo.org`, using the `inborutils` package. You can change the `path` option to where you want to download the archive (default is current directory '.'):
 
+```
+remotes::install_github("inbo/inborutils")
+inborutils::download_zenodo("10.5281/zenodo.13733183", path=".")
+```
+
+Store the path to the archive containing the reference databases and the bayesian models:
+
+```
+refdata_archive_path = "path/to/genomesizeRdata.tar.gz"
+```
+
+Read the example input file from the package. This example data is a subset of the dataset from [Labouyrie et al. 2023](https://doi.org/10.1038/s41467-023-37937-4):
+
+
+```
+example_input_file = system.file("extdata", "example_input.csv", package = "genomesizeR")
+```
+
+Load the package:
+
+```
+library(genomesizeR)
+```
+
+Run the main function to get the estimated genome sizes (with the default method which is the bayesian method):
 First, the genome sizes are predicted from the taxa:
 
 ```
@@ -33,7 +58,7 @@ First, the genome sizes are predicted from the taxa:
 
 Then, the results can be visualized using the plotting functions provided. This histogram shows the estimated genome sizes for each sample. 
 
-```{r, echo = FALSE}
+```{r, echo = FALSE, warning=F, message=F}
   library(genomesizeR)
   example_output_data = system.file("extdata", "results_bayesian.csv", package = "genomesizeR")
   refdata_archive_path = system.file("extdata", "genomesizeRdata.tar.gz", package = "genomesizeR")
@@ -68,6 +93,6 @@ This boxplot shows the estimated genome sizes for each sample:
 
 This tree shows the taxonomic relationships as well as the estimated genome sizes. The difference between the genome size distribution of bacteria (16S marker) and fungi (ITS marker) is visible.
 
-```{r, echo = TRUE, fig.width= 22, fig.height= 22}
+```{r, echo = TRUE, fig.width= 22, fig.height= 22, warning=F, message=F}
   plotted_df = plot_genome_size_tree(results, refdata_archive_path)
 ```
