+++ DESCRIPTION
@@ -11,7 +11,7 @@ Description: Genome size estimation from taxonomic data.
 License: GPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 Imports: pbapply, ggplot2, dplyr, lme4, merTools, brms, parallel, doParallel, ncbitax, ggtree, biomformat
 Remotes: bioc::ggtree, bioc::biomformat, github::raim/ncbitax
 Suggests: knitr, rmarkdown
