### 2025-10-01 4992eb9
Added links to repo and issues in doc and description
