### 2025-10-01 c929bf4
Input formats now better documented and with more robust checking
