### 2025-10-01 1cc7f0c
Plotting functions updated to not use deprecated aes_string() function
