### 2025-10-02 9964d00
Changed variable name for clarity
