### 2025-10-03 #8759 @celine-mercier
Dear reviewers, thank you for the helpful comments and suggestions. Addressing all comments:

@g-rppl :

1. _I am not sure whether the absence of tests fully aligns with JOSS criteria. In any case, including some basic functionality tests would strengthen the submission._

In the current submission, we provide a sample input file together with instructions on how to run all public functions included in the package. We hope that this current setup aligns with the “OK” category in the JOSS testing criteria, but if not, we will look into implementing an automated test suite. [JOSS review guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) state:

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.
>
> Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
> OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
> Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works

2. _I recommend moving the Example vignette to the top-level Get started link in the pkgdown site, as described [here](https://pkgdown.r-lib.org/reference/build_articles.html#get-started). This would help readers quickly access the main usage example and would make the Simple example section in the README unnecessary._

Done (https://github.com/ScionResearch/genomesizeR/commit/b7d94022485f75309ab0f6b7568886e0b3d7a8b1, https://scionresearch.github.io/genomesizeR/articles/genomesizeR.html) thank you for the suggestion.

3. _Consider adding links to the source repository and issue tracker in the DESCRIPTION file, and link them in the documentation as described [here](https://pkgdown.r-lib.org/articles/customise.html#navbar-heading)._

Done (https://github.com/ScionResearch/genomesizeR/commit/4992eb9ecca43ee0617256de73d81aa6385b4410, https://scionresearch.github.io/genomesizeR) thank you for the suggestion.

4. _Clarify the Bayesian method by elaborating the use of predictors, highlighting the hierarchical structure, and noting that the standard deviation is also estimated._

The Bayesian method is now described in more details (https://github.com/ScionResearch/genomesizeR/commit/3d73eb0059b17a878b542d8cfc52e1b81dac71ab). Please note that this more detailed description was in the original submission, but was edited to reduce the word count to around 1,000 during the pre-review, which means that the word count is now a bit too high again (1,222 words).

5. _Please consider citing brms as it seems to form a core part of the package._

Included in previous edit.

@BrandonEdwards :

6. _I have managed to get the package downloaded onto my computer with the help of the authors, so thank you. I think I might leave the Installation box unchecked for the time being, and would recommend that the authors provide additional installation options (as per https://github.com/ScionResearch/genomesizeR/issues/1) in the README._

Now providing 4 possible installation commands (https://github.com/ScionResearch/genomesizeR/commit/22b20134b4e34606c85f3a6815a57e048e7b95d6, https://github.com/ScionResearch/genomesizeR?tab=readme-ov-file#how-to-install).

7. _I also note that, in the example in the README, the first plotting function (i.e. plotted_df = plot_genome_size_histogram(results)) produces a note from ggplot2 that the use of aes_string() was deprecated in ggplot2 3.0.0. I don't think it's detrimental currently to the package, but keep in mind this may affect the package functionality in the future and would recommend fixing this soon._

The plotting functions were edited to not use the deprecated syntax (https://github.com/ScionResearch/genomesizeR/commit/1cc7f0c0741c6bd687d091c3a2c82bb9f5dfa572), however the `plot_tree()` function still triggers some warnings coming from the `ggtree` package. `ggtree` is actively maintained and the warnings should disappear once the `ggtree` authors update the package with the new syntaxes.

@msabrysarhan :

8. _Downloading and testing the reference data and the example csv file was not straight forward to me. It was not clear to me how to format the input file. I'd have preferred to have the function estimate_genome_size() accept taxids or vector of multiple taxids. However, this is very optional._

The input formats are now explained in more detail and with examples here: https://scionresearch.github.io/genomesizeR/articles/input_output.html#input including an example using a vector. The function documentation was also updated and the input checking made more robust in the code (https://github.com/ScionResearch/genomesizeR/commit/c929bf4c829dd8792541d81995bb9e7dbc59315e, https://scionresearch.github.io/genomesizeR/reference/estimate_genome_size.html).

9. _In the abstract, the authors mentioned "Several plotting functions ...". I only saw three plotting functions. I would recommend to remove "Several"._

Replaced "several" with "some" in the paper (https://github.com/ScionResearch/genomesizeR/commit/2567f26667aa67165a7ec771d3ccff88f267b6d1).

10. _I tried to run the function with other TaxIDs at species, genus and class level, but kept getting the same values.
Here is the taxa I tested:
TaxIDs: 162481, 1799530, 418497
Could you please double check for that? I expected to get still estimate at higher taxonomic levels._

We were not able to repeat this case:

`output = estimate_genome_size(c(162481, 1799530, 418497), 'inst/extdata/genomesizeRdata.tar.gz', format='vector')`

Although it is worth noting that some estimations can be identical in some cases when the closest known genome size(s) are shared, depending on the method. This typically happens more often when using the frequentist/lmm method.

Thank you again for the helpful comments, that we hopefully addressed adequately.
