### 2025-10-28 2940fe4
Updated the documentation for the latest version of the bayesian models
