### 2025-10-28 #8759 @celine-mercier
Dear reviewers,

While investigating we found an issue with version 1.0.2 of the bayesian models on zenodo (suboptimal behaviour when encountering new data). We re-optimised the estimation process of the Bayesian models, uploaded the new version to [zenodo](https://zenodo.org/records/17460393) and updated the documentation in consequence.

@msabrysarhan It looks to us like some older versions of the package and models were possibly being used. The only way we could replicate your output was by using the models of [v1.0.1](https://zenodo.org/records/13733781). This output also contains a column that was renamed in the later versions of the package. This could have happened from versions getting mixed within the R environment.

We now clearly tagged the latest package and model versions. Could you test that you are now getting all the correct outputs/behaviours:

```
# Update package and dependencies
remotes::install_github("https://github.com/ScionResearch/genomesizeR",
                                      dependencies = T, upgrade = 'always', force=T)

# Check version is correct
packageVersion('genomesizeR')   # Must be ‘1.0.0.2’

# Download new models and reference databases
inborutils::download_zenodo("10.5281/zenodo.13733183", path=".")

# Load package and test
library(genomesizeR)
output = estimate_genome_size(c(162481, 1799530, 418497), 'genomesizeRdata_v1.0.3.tar.gz', format='vector')

output[,1:4]
           taxa estimated_genome_size confidence_interval_lower confidence_interval_upper
query_1  162481              26739731                   6839614                  73074071
query_2 1799530              26779985                   6960343                  76837168
query_3  418497              27142733                   6350767                  85429833
```
