### 2025-12-17 76cab93
Added E2E tests for each method and doc for it
