+++ tests/testthat/setup.R
@@ -1,0 +1,4 @@
+# Test setup configuration
+# This file is automatically run by testthat before running tests
+
+# Configure any global test settings here if needed
