### 2025-12-18 #8759 @celine-mercier
* [x]  Double check Paper authors and affiliations (including ORCIDs)
* [x]  Make a tagged release of your software on Github, and list the version tag of the archived version here.
  [v1.0.2](https://github.com/ScionResearch/genomesizeR/releases/tag/v1.0.2)
  https://zenodo.org/records/17971801
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct title (should match the paper title).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct authors and affiliations (should match the paper authors including order).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct version (should match the above Github tag you created).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct ORCIDs (should match the paper authors).
  https://doi.org/10.5281/zenodo.17971801
