### 2025-12-21 eba28a9
Small paper edits for review
