### 2026-01-06 aa3bd32
Update mention and refs for No-U-Turn Sampler
