+++ paper/paper.bib
@@ -90,3 +90,21 @@
   encoding = {UTF-8},
 }
 
+@article{hoffman2014nuts,
+  author  = {Matthew D. Hoffman and Andrew Gelman},
+  title   = {The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo},
+  journal = {Journal of Machine Learning Research},
+  year    = {2014},
+  volume  = {15},
+  number  = {47},
+  pages   = {1593--1623},
+  url     = {http://jmlr.org/papers/v15/hoffman14a.html}
+}
+
+@manual{stanref2025,
+  author       = {{Stan Development Team}},
+  title        = {Stan Reference Manual, version 2.37.0},
+  year         = {2025},
+  note         = {Version 2.37.0},
+  url          = {https://mc-stan.org}
+}
