### 2026-01-16 5ec30a2
Update README
Added a section for citing information
